Create setup mechanism and optimize imports
This commit introduces a setup mechanism based on setuptools. As a result, it is now geared to being used as an installed package.
The imports have been modified/simplified to work with this new paradigm; import optimization has also been done in an effort to simplify and standardize the code structure.
The __load__opao.py files for configuring the path are no longer required and have been removed; the MKL availability check and additional functionality present in those files has been moved to the top level __init__.py.
The tutorials have also been modified to work with this new approach.
Instructions for installing this package have been added to the README file.